remove deprecated lib
